requirements.txt
fastapi
uvicorn
motor
requests
python-socketio
python-multipart